requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
deepsearch-glm==1.0.0
dill==0.3.9
docling==2.18.0
docling-core==2.17.1
docling-ibm-models==3.3.0
docling-parse==3.2.0
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.19
langchain-core==0.3.37
langchain-google-genai==2.0.11
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langsmith==0.3.8
latex2mathml==3.77.0
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
python-telegram-bot==21.10
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
semchunk==2.2.2
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
XlsxWriter==3.2.2
yarl==1.18.3
zstandard==0.23.0
